Generate i18n credits file when bumping translations
The i18n credits file was previously generated at build time, but doing
this when bumping translations makes more sense because this way
translators are added to the credits file when their translations are
added to the git repo (and not before that).

This change also removes the i18n credits file from .gitignore so it is
checked in the repo, and it is persisted. This will also make it easier
to track changes in the list of translators.

Also, an added benefit is that now everyone can build the extension.
This is because previously it was necessary to have a private API token
in order to generate the i18n credits file at build time (otherwise the
build failed).

Change-Id: Iecfd7ffd409aedc875f9d84cff0549f8d102d330