Open translation tab next to the current one
This new behavior will hopefully improve the user experience, by making
it identical to the behavior when clicking a link with the
target="_blank" attribute.

In particular, apart from open the tab next to the current one, when
closing it, the previous tab will be focused so the user can continue
reading the page.

Change-Id: I2c3629f7220d72928ee38481deb859c7f8bce591