build: build stable and beta channels with MV3
Manifest Version 2 will be deprecated in June 2024 in Chrome. This CL
switches the build system to release the beta and stable Chrome builds
with a MV3 build instead of a MV2 one.

It also removes the |web-ext lint| check in the lint job because
Chrome's MV3 builds are quite different from Firefox's MV3 builds (which
the extension doesn't provide) and this makes the |web-ext lint| check
fail. The lint job now only checks that the extension can be built.

Fixed: translateselectedtext:4
Change-Id: I2e0645aa1eac73b9e8bc7e7a551a625497127722